Refactoring in various files. Some double backslashes (\\) are replaced by single forward slash (/). Tried to find the bug where Avatar moves in the left direction at program start (in release mode only). Didn't find it though...

+++ CoreLib/src/GameStats.cpp
@@ -96,5 +96,5 @@ bool GameStats::isSuperMode()
 }
 float GameStats::superTimeRemaining()
 {
-	return 3 - m_superModeTimer->getElapsedTime();
+	return (float)(3 - m_superModeTimer->getElapsedTime());
 }
